fix latent warnings

+++ agent/zephyr/app/src/coap_help.h
@@ -8,6 +8,6 @@
 #define MAX_COAP_MSG_LEN 2048
 
 int send_coap_request(uint8_t* data, uint32_t len);
-int process_coap_reply(uint8_t* return_code, uint8_t* data, uint32_t* len);
+int process_coap_reply(uint8_t* return_code, uint8_t* data, uint16_t* len);
 
 #endif
